third_party/blink/web_tests/flag-specific/disable-layout-ng/fast/table/padding-height-and-override-height-expected.txt

PASS document.getElementById('cell-contents').offsetHeight is 78
FAIL document.getElementById('cell-contents').offsetWidth should be 88. Was 98.
PASS successfullyParsed is true

TEST COMPLETE
